3 opcode optimizations:
1)  OP_IMM+OP_POP is "peephole optimized" away at runtime.  This makes
verbdocs and other comments cheaper.

2)  OP_PUT_n+OP_POP is similarly optimized (PUT doesn't consume the
top value on the stack but it is often used that way in statements like
`var = expr;').  OP_G_PUT could use the same change but is rarely
executed.

3)  OP_PUT_n, OP_PUSH_n which used to be in an if/else in the default
case are split out into 32 cases each so the compiler can optimize it
for us.  These ops account for a large percentage of those executed.